pyproject.toml
[tool.poetry]
name = "2024django-reviews-backend"
version = "0.1.0"
description = ""
authors = ["jisoo Yeon <dydy11642@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"
pyjwt = "^2.8.0"
django-environ = "^0.11.2"
openai = "^1.25.1"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
